Moved RoleDomainMapper from Helper to Mapper namespace.
Moved \eZ\Publish\Core\Repository\Helper\RoleDomainMapper to \eZ\Publish\Core\Repository\Mapper\RoleDomainMapper.

Kept Helper\RoleDomainMapper for temporary BC with 1st party packages.
