+++ eZ/Publish/Core/Repository/Permission/PermissionResolver.php
@@ -16,7 +16,7 @@ use eZ\Publish\API\Repository\Values\User\LookupPolicyLimitations;
 use eZ\Publish\API\Repository\Values\User\UserReference as APIUserReference;
 use eZ\Publish\API\Repository\Values\ValueObject;
 use eZ\Publish\Core\Base\Exceptions\InvalidArgumentValue;
-use eZ\Publish\Core\Repository\Helper\RoleDomainMapper;
+use eZ\Publish\Core\Repository\Mapper\RoleDomainMapper;
 use eZ\Publish\SPI\Limitation\Target;
 use eZ\Publish\SPI\Limitation\TargetAwareType;
 use eZ\Publish\SPI\Limitation\Type as LimitationType;
@@ -35,7 +35,7 @@ class PermissionResolver implements PermissionResolverInterface
      */
     private $sudoNestingLevel = 0;
 
-    /** @var \eZ\Publish\Core\Repository\Helper\RoleDomainMapper */
+    /** @var \eZ\Publish\Core\Repository\Mapper\RoleDomainMapper */
     private $roleDomainMapper;
 
     /** @var \eZ\Publish\Core\Repository\Permission\LimitationService */
@@ -59,7 +59,7 @@ class PermissionResolver implements PermissionResolverInterface
     private $policyMap;
 
     /**
-     * @param \eZ\Publish\Core\Repository\Helper\RoleDomainMapper $roleDomainMapper
+     * @param \eZ\Publish\Core\Repository\Mapper\RoleDomainMapper $roleDomainMapper
      * @param \eZ\Publish\Core\Repository\Permission\LimitationService $limitationService
      * @param \eZ\Publish\SPI\Persistence\User\Handler $userHandler
      * @param \eZ\Publish\API\Repository\Values\User\UserReference $userReference
